requirements.txt
streamlit
pydicom 
numpy
Pillow